requirements.txt
Flask==2.3.3
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
openai-whisper==20231117
srt==3.5.3
anthropic==0.3.11
openai==0.28.1
tqdm==4.66.1
gunicorn==21.2.0